updating docstring
